sketching out a few things that would enable updating plugins on audio thread
just for demonstration purposes. I think it would make sense to repurpose & rename  ParamMgrNode / ParamMgrProcessor so that the ParamManager stuff is not the sole purpose of the node. The node would allow interactions with hosts on the audio thread and could also be the default place for devs to put custom code in case they don't need CompositeAudioNode